Fixing mpl product

+++ src/util/create_vmpl_sequence.hpp
@@ -198,7 +198,7 @@ struct vmpl_reduce_prod_impl<c,accu,stop,vmpl,true>
 template<typename vmpl>
 struct vmpl_reduce_prod
 {
-	typedef typename vmpl_reduce_prod_impl<0,1,vmpl::size::value-1,vmpl,false>::type type;
+	typedef typename vmpl_reduce_prod_impl<0,1,(int)vmpl::size::value-1,vmpl,false>::type type;
 };
 
 template<typename vmpl, int stop>
